Start up quiz page last known working state

+++ VikingQuiz.Api/Repositories/QuizRepository.cs
@@ -9,21 +9,22 @@ namespace VikingQuiz.Api.Repositories
     public class QuizRepository
     {
         private VikinQuizContext context;
-        public QuizRepository(VikinQuizContext context)
-        {
+        public QuizRepository(VikinQuizContext context) {
             this.context = context;
         }
 
-        public Quiz CreateQuiz(Quiz quiz)
-        {
+        public Quiz CreateQuiz(Quiz quiz) {
             context.Add(quiz);
             context.SaveChanges();
             return quiz;
         }
 
-        public Quiz UpdateQuiz(Quiz quiz)
-        {
-            Quiz foundQuiz = context.Quiz.Where(q => q.Id == quiz.Id).FirstOrDefault();
+        public Quiz UpdateQuiz(Quiz quiz) {
+            Quiz foundQuiz = context.Quiz.Where(dbQuiz => dbQuiz.Id == quiz.Id && dbQuiz.UserId == quiz.UserId).FirstOrDefault();
+            if (foundQuiz == null) {
+                return null;
+            }
+
             foundQuiz.Title = quiz.Title;
             foundQuiz.PictureUrl = quiz.PictureUrl;
             foundQuiz.UserId = quiz.UserId;
@@ -32,35 +33,49 @@ namespace VikingQuiz.Api.Repositories
             return quiz;
         }
 
-        public void DeleteQuiz(int id)
-        {
-            var games = context.Game.Where(g => g.QuizId == id).ToList();
+        public Quiz DeleteQuiz(int id, int userId) {
+            Quiz quiz = new Quiz {
+                Id = id
+            };
+
+
+            var quizMatchingIdAndPlayerId = context.Quiz.Where(dbQuiz => dbQuiz.Id == id && dbQuiz.UserId == userId).ToList();
+            if (quizMatchingIdAndPlayerId.Count == 0) {
+                return null;
+            }
+
+            /* >>>  the 2 lines below (the attempt to remove games) caused the server to crash 
+             *      (status 500) when trying to delete quizes which had questions assoicated with them 
+             *      (quizzes with no questons were deleted succesfully)
+            var games = context.Game.Where(x => x.QuizId == id).ToList();
             context.Game.RemoveRange(games);
-            var quizquestions = context.QuizQuestion.Where(qq => qq.QuizId == id).ToList();
-            context.QuizQuestion.RemoveRange(quizquestions);
-            var quizToRemove = context.Quiz.FirstOrDefault(q => q.Id == id);
-            context.Quiz.Remove(quizToRemove);
+            */
+
+            var quizQuestions = context.QuizQuestion.Where(question => question.QuizId == id).ToList();
+            context.QuizQuestion.RemoveRange(quizQuestions);
+
+            context.Quiz.Remove(quizMatchingIdAndPlayerId[0]);
             context.SaveChanges();
+
+            return quizMatchingIdAndPlayerId.FirstOrDefault();
         }
 
-        public Quiz GetQuizById(int id)
-        {
-            Quiz foundQuiz = context.Quiz.Where(q => q.Id == id)
-                .Select(q => new Quiz { Id = q.Id, Title = q.Title, UserId = q.UserId, PictureUrl = q.PictureUrl})
+        public Quiz GetQuizById(int id) {
+            Quiz foundQuiz = context.Quiz.Where(dbQuiz => dbQuiz.Id == id)
+                .Select(quiz => new Quiz { Id = quiz.Id, Title = quiz.Title, UserId = quiz.UserId, PictureUrl = quiz.PictureUrl })
                 .FirstOrDefault();
 
             return foundQuiz;
         }
 
-        public IEnumerable<Quiz> GetQuizByUserId(int id)
-        {
+        public List<Quiz> GetQuizByUserId(int id) {
             return context.Quiz.Where(q => q.UserId == id).OrderByDescending(q => q.LastModified).ToList();
 
         }
 
-        public IEnumerable<Quiz> GetAll()
-        {
-            return context.Quiz.ToList();
+        public IEnumerable<Quiz> GetAll(int userId) {
+            var quizzesOfTheUser = context.Quiz.Where(dbQuiz => dbQuiz.UserId == userId).ToList();
+            return quizzesOfTheUser;
         }
     }
 }
